[#4] more linting passing.
