+++ nlp4all/__init__.py
@@ -8,7 +8,7 @@ from flask_sqlalchemy import SQLAlchemy
 from flask_bcrypt import Bcrypt
 from flask_login import LoginManager
 from flask_mail import Mail
-from nlp4all import routes
+
 
 app = Flask(__name__)
 app.config["SECRET_KEY"] = "5791628bb0b13ce0c676dfde280ba245"
