+++ nlp4all/routes.py
@@ -48,10 +48,25 @@ import datetime
 import json
 import ast
 from sqlalchemy.orm.attributes import flag_modified
-from nlp4all.utils import get_user_projects, get_user_project_analyses
+from nlp4all.utils import get_user_projects
 import operator
 import re
 
+# @TODO: refactor this, i moved this function here
+# because it was in utils, which imported models which imported utils
+def get_user_project_analyses(a_user, a_project):
+    analyses = BayesianAnalysis.query.filter_by(project=a_project.id)
+    if current_user.admin:
+        return analyses
+    else:
+        return [a for a in analyses if a.shared or a.shared_model or a.user == current_user.id]
+    # if a_user.admin:
+    #         return(BayesianAnalysis.query.filter_by(project=a_project.id).all())
+    # else:
+    #         analyses = []
+    #         all_project_analyses = BayesianAnalysis.query.filter_by(project=a_project.id)
+    # return [a for a in all_project_analyses if a.shared or a.user ==
+    # a_user.id]
 
 @app.route("/")
 @app.route("/home")
