+++ run.py
@@ -1,3 +1,6 @@
+"""
+nlp4all flask app runner.
+"""
 from nlp4all import app
 
 if __name__ == "__main__":
